+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Warning
+- On v0.2.2 release we added the `obscure` argument to the rclone remote generation process. This means that the passwords and api keys will be encrypted in the rclone config file. If you have remotes like `crypt` or similar to required keys, you need to re-create or update them. Otherwise keys will not be encrypted. (#90)
+
 ### Added
 - Windows version of headless builds now available.
 - Homebrew tap for macOS added.
@@ -18,8 +21,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Fixed
 - App UI now update properly after values changed.
 - RPM dependencies fixed.
+- CLI `tray` argument fixed.
 
 ## [v0.2.2] - 2026-03-11
+
+### Warning
+- While this in this release we added the `obscure` argument to the rclone remote generation process. This means that the passwords and api keys will be encrypted in the rclone config file. If you have remotes like `crypt` or similar to required keys, you need to re-create or update them. Otherwise keys will not be encrypted. (#90)
 
 ### Added
 - **Language**: Added Spanish language support. (Thanks to @dikler!)